requirements.txt
python-telegram-bot==20.0
textblob
requests
Pillow
pytesseract
google-generativeai
dotenv
os
emoji
uuid
json
subprocess
PyPDF2
pdf2image
pytesseract
tiktoken 
datetime
urllib